3 Reasoning Act Observe/lab/extensions/requirements.txt
google-generativeai>=0.5.0
requests>=2.32.0
openai>=1.0.0
